+++ uninstall.sh
@@ -75,6 +75,7 @@ rm -f \
     "$configurator_directory/window-layouts-configurator-service" \
     "$configurator_directory/configurator.py" \
     "$configurator_directory/updater.py" \
+    "$configurator_directory/layout_transfer.py" \
     "$configurator_directory/VERSION" \
     "$configurator_directory/window-layouts.svg" \
     "$configurator_service"